jpeg photo reverse when filtering

+++ ViperPicturesDownloader/Services/FileProvider/FileProvider.swift
@@ -5,12 +5,9 @@ final class FileProvider: FileProviderProtocol {
 	let fileManager = FileManager()
 	let tempDirectory = NSTemporaryDirectory()
 	let limitedSizeFile = 10000000000
-	static var work = 0
 
 	/// Записывает данные в файл
 	func writeToFile(data: Data, path: String){
-		FileProvider.work = FileProvider.work + 1
-		print("WORKTT = \(FileProvider.work)")
 		do {
 			try data.write(to: URL(fileURLWithPath: path))
 			print("File text.txt created at temp directory")
@@ -20,8 +17,8 @@ final class FileProvider: FileProviderProtocol {
 	}
 	/// Проверяет директорию на существования файла
 	func checkDirectory(nameFile: String) ->Bool {
-		do {
-			let filesInDirectory = try fileManager.contentsOfDirectory(atPath: tempDirectory)
+//		do {
+			guard let filesInDirectory = getAllFilesFromDirectory() else { return false}
 			for file in filesInDirectory {
 				if file == nameFile{
 					print("\(nameFile) found")
@@ -29,9 +26,9 @@ final class FileProvider: FileProviderProtocol {
 				}
 			}
 			print("File not found")
-		   } catch let error as NSError {
-			   print(error)
-		   }
+//		   } catch let error as NSError {
+//			   print(error)
+//		   }
 		   return false
 	   }
 	/// Считывает данные с файла
@@ -51,17 +48,17 @@ final class FileProvider: FileProviderProtocol {
 	}
 	/// Удаляет ненужные файлы
 	func removeFile(nameFile: String, before date: Date? = Date()) {
+		let path = getPath(nameFile: nameFile)
         do {
-			let path = getPath(nameFile: nameFile)
 			let dateFile = try fileManager.attributesOfItem(atPath: path)[FileAttributeKey.creationDate] as? Date
 			let sizeFile = try fileManager.attributesOfItem(atPath: path)[FileAttributeKey.size] as? Int
 			if let date = date, let dataFile = dateFile, let sizeFile = sizeFile {
 				if date > dataFile || sizeFile > limitedSizeFile {
-					try fileManager.removeItem(atPath: path)
+					removeItem(path: path)
 					print("file deleted")
 				}
 			} else {
-				try fileManager.removeItem(atPath: path)
+				removeItem(path: path)
 				print("file deleted")
 			}
         } catch let error as NSError {
@@ -70,26 +67,55 @@ final class FileProvider: FileProviderProtocol {
 	}
 	/// Удаляет все файлы
 	func removeAllFiles(before date: Date? = Date()) {
-		do {
-			let filesInDirectory = try fileManager.contentsOfDirectory(atPath: tempDirectory)
-			for file in filesInDirectory {
-				removeFile(nameFile: file, before: date)
-			}
-		} catch let error as NSError {
-            print("error occured while deleting file: \(error.localizedDescription)")
-        }
+//		do {
+		guard let filesInDirectory = getAllFilesFromDirectory() else {return}
+		for file in filesInDirectory {
+			removeFile(nameFile: file, before: date)
+		}
+//		} catch let error as NSError {
+//            print("error occured while deleting file: \(error.localizedDescription)")
+//        }
 	}
+//	func removeFilesWithType() {
+//		let files = getAllFilesFromDirectory()
+//		do {
+//			let filesInDirectory = try fileManager.contentsOfDirectory(atPath: tempDirectory)
+//			for file in filesInDirectory {
+//				if file.hasSuffix(".jpeg") || file.hasSuffix(".png") {
+//					let path = getPath(nameFile: file)
+//					try fileManager.removeItem(atPath: path)
+//				}
+//			} //try fileManager.attributesOfItem(atPath: path)[FileAttributeKey.size] as? Int
+//		} catch let error as NSError {
+//            print("error occured while deleting file: \(error.localizedDescription)")
+//        }
+//	}
+
 	func removeFilesWithType() {
+		guard let filesInDirectory = getAllFilesFromDirectory() else {return}
+		for file in filesInDirectory {
+			if file.hasSuffix(".jpeg") || file.hasSuffix(".png") {
+				let path = getPath(nameFile: file)
+				removeItem(path: path)
+			}
+		}
+	}
+
+	private func removeItem(path: String) {
 		do {
-			let filesInDirectory = try fileManager.contentsOfDirectory(atPath: tempDirectory)
-			for file in filesInDirectory {
-				if file.hasSuffix(".jpeg") {
-					let path = getPath(nameFile: file)
-					try fileManager.removeItem(atPath: path)
-				}
-			} //try fileManager.attributesOfItem(atPath: path)[FileAttributeKey.size] as? Int
-		} catch let error as NSError {
-            print("error occured while deleting file: \(error.localizedDescription)")
-        }
+			try fileManager.removeItem(atPath: path)
+		} catch {
+			print(error)
+		}
+	}
+
+	private func getAllFilesFromDirectory(directory: String = NSTemporaryDirectory())-> [String]? {
+		do {
+			let filesInDirectory = try fileManager.contentsOfDirectory(atPath: directory)
+			return filesInDirectory
+		} catch {
+			print(error)
+		}
+		return nil
 	}
 }
